Adding much more robust newline trimming on outputs
